// wren-ui/src/types/apiHistory.ts
export enum ApiType {
  ASK = 'ASK',
  CREATE_CONNECTOR = 'CREATE_CONNECTOR',
  CREATE_INSTRUCTION = 'CREATE_INSTRUCTION',
  CREATE_KNOWLEDGE_BASE = 'CREATE_KNOWLEDGE_BASE',
  CREATE_SKILL = 'CREATE_SKILL',
  CREATE_SQL_PAIR = 'CREATE_SQL_PAIR',
  CREATE_VIEW = 'CREATE_VIEW',
  DELETE_CONNECTOR = 'DELETE_CONNECTOR',
  DELETE_INSTRUCTION = 'DELETE_INSTRUCTION',
  DELETE_SKILL = 'DELETE_SKILL',
  DELETE_SQL_PAIR = 'DELETE_SQL_PAIR',
  DELETE_THREAD = 'DELETE_THREAD',
  DELETE_VIEW = 'DELETE_VIEW',
  GENERATE_SQL = 'GENERATE_SQL',
  GENERATE_SUMMARY = 'GENERATE_SUMMARY',
  GENERATE_VEGA_CHART = 'GENERATE_VEGA_CHART',
  GET_API_HISTORY = 'GET_API_HISTORY',
  GET_CONNECTORS = 'GET_CONNECTORS',
  GET_INSTRUCTIONS = 'GET_INSTRUCTIONS',
  GET_KNOWLEDGE_BASES = 'GET_KNOWLEDGE_BASES',
  GET_MODELS = 'GET_MODELS',
  GET_SKILLS = 'GET_SKILLS',
  GET_SQL_PAIRS = 'GET_SQL_PAIRS',
  GET_THREADS = 'GET_THREADS',
  PREVIEW_MODEL_DATA = 'PREVIEW_MODEL_DATA',
  PREVIEW_VIEW_DATA = 'PREVIEW_VIEW_DATA',
  REENCRYPT_SECRETS = 'REENCRYPT_SECRETS',
  RUN_SQL = 'RUN_SQL',
  STREAM_ASK = 'STREAM_ASK',
  STREAM_GENERATE_SQL = 'STREAM_GENERATE_SQL',
  TEST_CONNECTOR = 'TEST_CONNECTOR',
  UPDATE_CONNECTOR = 'UPDATE_CONNECTOR',
  UPDATE_INSTRUCTION = 'UPDATE_INSTRUCTION',
  UPDATE_KNOWLEDGE_BASE = 'UPDATE_KNOWLEDGE_BASE',
  UPDATE_SKILL = 'UPDATE_SKILL',
  UPDATE_SQL_PAIR = 'UPDATE_SQL_PAIR',
  UPDATE_THREAD = 'UPDATE_THREAD',
}

export type ApiHistoryResponse = {
  apiType: ApiType;
  createdAt: string;
  durationMs?: number | null;
  headers?: any;
  id: string;
  requestPayload?: any;
  responsePayload?: any;
  statusCode?: number | null;
  threadId?: string | null;
  updatedAt: string;
};